lab 3

+++ simulation/simulation_facade.h
@@ -5,8 +5,8 @@
 #include "../adt/matrix/matrix.h"
 #include "../adt/dstring/dstring.h"
 
-typedef Matrix *(*Function_u)(int);
-typedef Matrix *(*Function_b)(double, int);
+typedef Matrix *(*Function_u)(double);
+typedef Matrix *(*Function_b)(double, double);
 
 // Methods
 int start_simulation(int simulation_time, Function_u u_t, Function_b b_t);
